try making incremental interpreter use builtins
