+++ core/incremental/testdata/buildBinExpr.dd.txt
@@ -4,10 +4,9 @@ bar{a: A, b: B} :-
 ----
 application/graphviz
 digraph G {
-  // unmapped: bar [0, 1, 2, bar]
   "foo" [fillcolor="darksalmon" label="foo:  []" shape="box" style="filled"];
   "0" [fillcolor="darkseagreen2" label="0: Match(foo{a: A, b: B}; {}) [A-B]" shape="box" style="filled"];
-  "1" [fillcolor="darkseagreen2" label="1: Match(base.neq{a: A, b: B}; {}) [A-B]" shape="box" style="filled"];
+  "1" [fillcolor="" label="1: Builtin(base.neq{a: A, b: B}) [A-B]" shape="box" style="filled"];
   "2" [fillcolor="thistle" label="2: Join(A, B) []" shape="box" style="filled"];
   "bar" [fillcolor="lightblue" label="bar: Subst({a: A, b: B}) []" shape="box" style="filled"];
   "foo" -> "0" [];
