+++ core/incremental/testdata/family.dd.txt
@@ -17,22 +17,22 @@ parent: [parent{child: "Pete", parent: "Mary"}; {C: "Pete", P: "Mary"}]
 3: [parent{child: "Pete", parent: "Mary"}; {A: "Pete", B: "Mary"}]
 6: [parent{child: "Pete", parent: "Mary"}; {A: "Pete", B: "Mary"}]
 7: [parent{child: "Pete", parent: "Mary"}; {B: "Pete", C: "Mary"}]
-19: [parent{child: "Pete", parent: "Mary"}; {L: "Pete", P1: "Mary"}]
-22: [parent{child: "Pete", parent: "Mary"}; {P2: "Mary", R: "Pete"}]
-24: [parent{child: "Pete", parent: "Mary"}; {C: "Pete", P: "Mary"}]
-27: [parent{child: "Pete", parent: "Mary"}; {C: "Pete", P: "Mary"}]
-37: [parent{child: "Pete", parent: "Mary"}; {X: "Pete", Z: "Mary"}]
-38: [parent{child: "Pete", parent: "Mary"}; {X: "Pete", Y: "Mary"}]
+20: [parent{child: "Pete", parent: "Mary"}; {L: "Pete", P1: "Mary"}]
+23: [parent{child: "Pete", parent: "Mary"}; {P2: "Mary", R: "Pete"}]
+25: [parent{child: "Pete", parent: "Mary"}; {C: "Pete", P: "Mary"}]
+28: [parent{child: "Pete", parent: "Mary"}; {C: "Pete", P: "Mary"}]
+38: [parent{child: "Pete", parent: "Mary"}; {X: "Pete", Z: "Mary"}]
+39: [parent{child: "Pete", parent: "Mary"}; {X: "Pete", Y: "Mary"}]
 2: []
 5: []
 8: []
 8: []
-21: []
-23: []
-26: []
-29: []
-36: [parent{child: "Pete", parent: "Mary"}; {X: "Pete", Z: "Mary"}]
-40: []
+22: []
+24: []
+27: []
+30: []
+37: [parent{child: "Pete", parent: "Mary"}; {X: "Pete", Z: "Mary"}]
+41: []
 ancestor: [ancestor{an: "Mary", desc: "Pete"}; {X: "Pete", Z: "Mary"}]
-39: [ancestor{an: "Mary", desc: "Pete"}; {Y: "Pete", Z: "Mary"}]
-40: []
+40: [ancestor{an: "Mary", desc: "Pete"}; {Y: "Pete", Z: "Mary"}]
+41: []
