+++ core/simple/simpleEvaluate.ts
@@ -27,6 +27,7 @@ import { BUILTINS } from "../builtins";
 import { DB } from "./types";
 import { AGGREGATIONS } from "../aggregations";
 import { getForScope } from "./indexes";
+import { evalBuiltin } from "../evalBuiltin";
 
 export function evaluate(db: DB, term: Term): Res[] {
   const cache: Cache = {};
@@ -131,18 +132,7 @@ function doEvaluate(
           }
           const builtin = BUILTINS[term.relation];
           if (builtin) {
-            const substituted = substitute(term, scope) as Rec;
-            const records = builtin(substituted);
-            // console.log({ substituted: ppt(substituted), res: res.map(ppr) });
-            const results = records.map(
-              (rec): Res => ({
-                term: rec,
-                bindings: unify(scope, rec, term),
-                trace: { type: "BaseFactTrace" }, // TODO: BuiltinTrace?
-              })
-            );
-            // console.log(results.map(ppr));
-            return results;
+            return evalBuiltin(term, scope);
           }
           const rule = db.rules.get(term.relation);
           if (rule) {
